Django/Ayo/requirements.txt
asgiref==3.3.4
autopep8==1.5.6
certifi==2020.12.5
chardet==4.0.0
coverage==5.5
Django==3.2
django-cors-headers==3.7.0
django-mysql==3.11.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
idna==2.9
mysqlclient==2.0.3
pika==1.2.0
Pillow==8.2.0
pycodestyle==2.7.0
pyflakes==2.3.1
PyJWT==2.0.1
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.4
